供应链管理项目初始化

+++ bio-drqi-bsm/bio-drqi-bsm-service/src/main/java/com/bio/drqi/bsm/service/impl/BmsOrderDetailServiceImpl.java
@@ -119,7 +119,7 @@ public class BmsOrderDetailServiceImpl implements BmsOrderDetailService {
         if (bmsOrderDetailTb == null) {
             throw new BusinessException("找不到此订单");
         }
-        if (bmsOrderDetailTb.getReceiveNumber() != bmsOrderDetailTb.getPurchaseNumber()) {
+        if (bmsOrderDetailTb.getReceiveNumber() != bmsOrderDetailTb.getPurchaseNumber().intValue()) {
             throw new BusinessException("耗材未全部到货");
         }
